moved helpers to mixin and abstracted build method
